Parsing seems reasonably solid at this point 900 font files to test on Moving some tables to todo folder until I can find any documentation on them

+++ NewFontParser/Tables/Gpos/ScriptTable.cs
@@ -10,7 +10,7 @@ namespace NewFontParser.Tables.Gpos
 
         public LangSysRecord[] ScriptRecords { get; }
 
-        public ScriptTable(BigEndianReader reader, ushort offset)
+        public ScriptTable(BigEndianReader reader, long offset)
         {
             reader.Seek(offset);
